Fixed feedback bug and intro message bug
