+++ app/frontend-healthier-me/src/app/components/navbar/navbar-profile-links/navbar-profile-link.component.html
@@ -12,7 +12,7 @@ tw-cursor-pointer
 {{ isActive ? 'tw-bg-gray-100' : 'tw-bg-white hover:bg-gray-50' }}
 "
 >
-  <img *ngIf="profile.id === 'general'" ngSrc="/app/profiles/general.svg" height="40" width="40" class="tw-rounded-full tw-w-[40px] tw-h-[40px]" alt="" />
+  <img *ngIf="profile.id === 'general'" ngSrc="/app/profiles/general.svg" height="40" width="40" class="tw-rounded-full tw-w-[40px] tw-h-[40px]" alt="" />  
   <div *ngIf="profile.id !== 'general'" class="tw-w-[40px] tw-h-[40px] tw-rounded-full tw-bg-gray-300"></div>
   <div class="tw-flex tw-flex-col tw-justify-start tw-items-stretch">
     <span class="tw-capitalize">{{ profile.name }}</span>
