+++ app/frontend-healthier-me/src/app/components/voice-mobile/voice-annotation/voice-annotation.component.html
@@ -5,7 +5,8 @@
 </div>
 
 <!-- Feedback Modal -->
-<p-dialog [(visible)]="displayModal" [modal]="true" [closable]="true" header="Feedback">
+<p-dialog [(visible)]="displayNegativeModal" [modal]="true" [closable]="true" header="Feedback">
+  <p>We apologise for the negative experience. Please let us know how we can improve by leaving your feedback below: </p>
   <!-- Static Category Buttons -->
   <div class="tw-flex tw-justify-between tw-mt-4 tw-gap-2 tw-mb-2">
     <p-button
@@ -24,6 +25,63 @@
       label="Does Not Answer Question"
       (click)="selectCategory('Does Not Answer Question')"
       [severity]="selectedCategory === 'Does Not Answer Question' ? 'success' : 'secondary'"
+    >
+    </p-button>
+    <p-button
+      label="Irrelevant Tone"
+      (click)="selectCategory('Irrelevant Tone')"
+      [severity]="selectedCategory === 'Irrelevant Tone' ? 'success' : 'secondary'"
+    >
+    </p-button>
+    <p-button label="Others" (click)="selectCategory('Others')" [severity]="selectedCategory === 'Others' ? 'success' : 'secondary'"> </p-button>
+  </div>
+
+  <div class="field">
+    <label for="feedback">Remarks:</label>
+    <textarea
+      id="feedback"
+      pInputTextarea
+      [(ngModel)]="remarks"
+      placeholder="Please leave your feedback here. "
+      rows="5"
+      cols="30"
+      class="tw-w-full tw-p-2 tw-rounded-lg tw-border-2"
+    ></textarea>
+  </div>
+
+  <div class="tw-mt-3">
+    <button pButton label="Submit" icon="pi pi-check" (click)="submitFeedback()"></button>
+  </div>
+</p-dialog>
+
+<!-- Feedback Modal -->
+<p-dialog [(visible)]="displayPositiveModal" [modal]="true" [closable]="true" header="Feedback">
+  <p>We are happy that you had a good experience! Please tell us what we did well by leaving your feedback below: </p>
+  <!-- Static Category Buttons -->
+  <div class="tw-flex tw-justify-between tw-mt-4 tw-gap-2 tw-mb-2">
+    <p-button
+      label="Relevant Sources"
+      (click)="selectCategory('Outdated Sources')"
+      [severity]="selectedCategory === 'Relevant Sources' ? 'success' : 'secondary'"
+    >
+    </p-button>
+    <p-button
+      label="Factually Correct"
+      (click)="selectCategory('Factually Correct')"
+      [severity]="selectedCategory === 'Factually Correct' ? 'success' : 'secondary'"
+    >
+    </p-button>
+    <p-button
+      label="Straightforward Answer"
+      (click)="selectCategory('Straightforward Answer')"
+      [severity]="selectedCategory === 'Straightforward Answer' ? 'success' : 'secondary'"
+    >
+    </p-button>
+
+    <p-button
+    label="Relevant Tone"
+    (click)="selectCategory('Relevant Tone')"
+    [severity]="selectedCategory === 'Relevant Tone' ? 'success' : 'secondary'"
     >
     </p-button>
     <p-button label="Others" (click)="selectCategory('Others')" [severity]="selectedCategory === 'Others' ? 'success' : 'secondary'"> </p-button>
