+++ app/frontend-healthier-me/src/app/services/chat-message/chat-message.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from "@angular/core";
 import { NgxIndexedDBService } from "ngx-indexed-db";
 import { BehaviorSubject, firstValueFrom } from "rxjs";
-import { Message } from "../../types/message.type";
+import { Message, MessageRole } from "../../types/message.type";
 
 @Injectable({
   providedIn: "root"
@@ -24,7 +24,23 @@ export class ChatMessageService {
   async load(profileId: string): Promise<BehaviorSubject<Message[]>> {
     this.$currentProfileId.next(profileId);
 
+    const introMessage: Message = {
+      id: "intro-message", // Assign a unique ID for the intro message
+      profile_id: profileId,
+      role: MessageRole.Assistant,
+      message:
+        "Hello! I’m HealthierME, a friendly conversational assistant to help you learn more about health information and programmes in Singapore. Tap the voice button to ask a question!",
+      timestamp: Date.now(),
+      sources: []
+    };
+
     const messages = await firstValueFrom<Message[] | undefined>(this.indexedStore.getAllByIndex("messages", "profile_id", IDBKeyRange.only(profileId)));
+    // if no messages exist, put in intro message into array without updating the store
+    if (!messages || messages.length === 0) {
+      this.$messages = new BehaviorSubject<Message[]>([introMessage]);
+      return this.$messages;
+    }
+
     this.$messages = new BehaviorSubject<Message[]>(messages?.sort((a, b) => a.timestamp - b.timestamp) || []);
 
     return this.$messages;
